[FIX/#44] Int -> Long 변경

+++ app/src/main/java/com/napzak/market/presentation/explore/ExploreViewModel.kt
@@ -271,7 +271,7 @@ class ExploreViewModel @Inject constructor(
         getExploreProductInformation()
     }
 
-    fun updateItemLikeButton(productId: Int) {
+    fun updateItemLikeButton(productId: Long) {
         /* TODO: 좋아요 API 연결 및 기능 연결 */
     }
